Ajout de la récupération des données depuis une API et suppression de l'ancien fichier API
