+++ app/lib/api.tsx
@@ -1,0 +1,15 @@
+const fetchDatabase = async () => {
+  try {
+    const response = await fetch('https://opentdb.com/api.php?amount=10');
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+    const data = await response.json();
+    return data;
+  } catch (error) {
+    console.error('Error fetching data:', error);
+    return null;
+  }
+};
+
+export default fetchDatabase;
