+++ app/page.tsx
@@ -1,3 +1,24 @@
+'use client';
+import { useEffect, useState } from 'react';
+import fetchDatabase from './lib/api';
+
 export default function Page() {
-  return <h1>Hello, Next.js!</h1>
+  const [data, setData] = useState(null);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      const data = await fetchDatabase();
+      setData(data);
+    };
+
+    setTimeout(() => {
+      fetchData();
+    }, 1000);
+  }, []);
+
+  return (
+    <div>
+      {data ? <pre>{JSON.stringify(data, null, 2)}</pre> : <p>Loading...</p>}
+    </div>
+  );
 }
