--- app/pages/api.tsx
@@ -1,15 +1,0 @@
-const fetchDatabase = async () => {
-    try {
-        const response = await fetch('https://opentdb.com/api.php?amount=10');
-        if (!response.ok) {
-            throw new Error(`HTTP error! status: ${response.status}`);
-        }
-        const data = await response.json();
-        return data;
-    }
-    catch (error) {
-        console.log("Error fetching database:", error);
-    }
-}
-
-fetchDatabase();
